mkt-simulator/requirements.txt
google-cloud-pubsub==2.19.0